Fixed video slides implementation to support videos taking up entire slide.
Previously, height was the same for presentations with and without controls.
This left a large amount of whitespace at the bottom when there were no controls present.

Now, render.js checks to see if the controls are present or not and sets a video_height attribute.
This attribute is then passed to mustache where it is used by the video template.

I also realized that templates (except layout.mustache) could not tell if controls were present or not.
I added a single line to the rendering process to fix this.

Signed-off-by: Nicholas Bevacqua <videogamerz400@gmail.com>