Add paging.
A really useful feature I've missed from other presenting tools is the ability
to type a page number, press enter and instantly jump to that page.

I've added some really simple code to perform this. A typed number followed by
an enter will jump to that page. Enter with no number beforehand will still
go to the next page.

Escape will clear the number in the buffer if a mistake should be made with
this.